Allow for a single statement to fail tranformation
If a single statement fails (due to a course or discussion that it refers to having been deleted, for example), then no statements are sent. This same statement comes up first next time and thus the plugin is effectively disabled.

This change allows for a single statement to fail without stopping all others.