PR-timezonefix: Set the timezone back to site default after the manual Europe/London change.
Previously this set the timezone to Europe/London outside of this function, which was causing problems
elsewhere in the system, as once the plugin settings were all loaded by Moodle, the timezone was changing
from the site default. So this change puts it inside the function and then sets it back to the site default
when we are done.